FIX Error in browser console, when loading the service-worker.js for svelte
In Svelte and SvelteKit there is no index.html due to the routing. So / is the entrypoint to the app. I think that this should be reflected in the createHandlerBoundRoute() function. So I changed it to / and the uncaugth Error from the browsers console does no longer show up. Precaching works correct and the registration of the sevice woker is now possible without moving to state "redundant" any more.
I think that it was just a typo, because in the .ts files of template-custom-svelte-kit it was already set to "/".